인증, 권한 처리

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -40,7 +41,15 @@ public class BoardController {
 
     @GetMapping("/board/{id}")
     public String detail(@PathVariable int id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+        boolean isOwner = false;
+        if (sessionUser != null){
+            if (sessionUser.getId() == board.getUser().getId()) {
+                isOwner = true;
+            }
+        }
+        request.setAttribute("isOwner", isOwner);
         request.setAttribute("board", board);
 
         return "/board/detail";
@@ -56,6 +65,11 @@ public class BoardController {
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable int id, BoardRequest.UpdateDTO requestDTO) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+        if(sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 수정할 권한이 없습니다.");
+        }
         boardRepository.updateById(id, requestDTO);
 
         return "redirect:/board/" + id;
@@ -63,6 +77,11 @@ public class BoardController {
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable int id) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+        if(sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 삭제할 권한이 없습니다.");
+        }
         boardRepository.deleteById(id);
 
         return "redirect:/";
